Fix [GH23] 'Not working ... because of SMC event'.
Fix [GH23] 'Not working on Macbook 5,5 because of SMC event'.

On Macbook 5.5 (maybe others) the SMC is constantly sending events,
probably from the light sensor or something so we want to ignore this.

The SMC is the system management controller. It's responsible for a
number of processes, including the cooling fans, keyboard, and LED
lights.

Thanks to 'robinmayol' for providing the solution.