Startup delay set to 2 if < 0 after action.
If the startup delay is below 1 minute (0 or negative), then it is set
to 2 minutes after the action command is executed once.

This is to prevent the action command to be repeatedly executed very
quickly.  If the command is for instance, 'Sleep', without this fix /
safeguard the computer would always go to sleep immediately after waking
up, giving no chance to the user to take control of it.